# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user-achievements/create/{userId}/{achievementId}": {
      "post": {
        "operationId": "UserAchievementController_createUserAchievement",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "achievementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/user-achievements/user/{userId}": {
      "get": {
        "operationId": "UserAchievementController_getUserAchievementsByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Ft Transcendence API",
    "description": "The Ft Transcendence API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}